requirements.txt
# Core dependencies
pyserial
numpy
ahrs
flask
flask-socketio
pigpio
scipy
pyyaml

# Development/testing dependencies
pytest
pytest-cov
mypy